ci: try to fix build in Vercel with Nodejs v18
