+++ .github/actions/prepare/action.yml
@@ -10,9 +10,7 @@ runs:
       # https://github.com/actions/setup-node
       uses: actions/setup-node@v3
       with:
-        # use node 16 due to avoid umi3(webpack4) building error:
-        # `Error: error:0308010C:digital envelope routines::unsupported`
-        node-version: 16.x
+        node-version: 18.x
         registry-url: https://registry.npmjs.org
 
     - name: Environment variables
