+++ examples/umi3/.umirc.ts
@@ -1,4 +1,5 @@
 // https://umijs.org/config/
+import path from 'path'
 import { defineConfig } from 'umi'
 
 const isDev = process.env.NODE_ENV === 'development'
@@ -20,13 +21,18 @@ export default defineConfig({
   extraBabelPlugins: [
     [
       // https://github.com/emotion-js/emotion/tree/main/packages/babel-plugin#options
-      '@emotion/babel-plugin',
+      require.resolve('@emotion/babel-plugin'),
       {
         sourceMap: true,
         autoLabel: 'always',
         labelFormat: '[local]',
       },
     ],
+  ],
+  extraBabelIncludes: [
+    // for support `?.` (optional-chaining) and `??` (nullish-coalescing-operator) in umi3 webpack4
+    'react-dev-inspector',
+    path.join(__dirname, '../../packages/inspector/es'),
   ],
 
   plugins: [
