+++ examples/umi3/package.json
@@ -15,9 +15,10 @@
   },
   "scripts": {
     "clean": "rimraf dist src/.umi src/.umi-production",
-    "dev": "npm run clean && COMPRESS=none HOST=0.0.0.0 MFSU_AD=none umi dev",
+    "umi": "NODE_OPTIONS=\"${NODE_OPTIONS} --openssl-legacy-provider\" umi",
+    "dev": "npm run clean && COMPRESS=none HOST=0.0.0.0 MFSU_AD=none npm run umi dev",
     "build": "NODE_ENV=production npm-run-all clean build:*",
-    "build:umi": "COMPRESS=none SPEED_MEASURE=CONSOLE umi build"
+    "build:umi": "COMPRESS=none SPEED_MEASURE=CONSOLE npm run umi build"
   },
   "dependencies": {
     "@emotion/css": "11.1.3",
